# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Match-Appi API",
        "description": "API REST para la porra del Mundial 2026",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/champion-predictions": {
            "get": {
                "summary": "Lista predicciones de campeón (admin: todas, user: la suya)",
                "operationId": "8910c808a3aa1e56507a8005c90d4cd1",
                "responses": {
                    "200": {
                        "description": "Lista de predicciones de campeón"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Crear predicción de campeón (solo user, durante fase de grupos)",
                "operationId": "03fc13b0dfe7262802f0a275ad2145b1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "team_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Predicción de campeón creada"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación o fase de grupos terminada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/champion-predictions/{championPrediction}": {
            "put": {
                "summary": "Actualizar predicción de campeón (solo propietario, durante fase de grupos)",
                "operationId": "b336d33fe058aab6333a881985aa9d1b",
                "parameters": [
                    {
                        "name": "championPrediction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "team_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Predicción de campeón actualizada"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Fase de grupos terminada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Eliminar predicción de campeón (propietario o admin)",
                "operationId": "61ad2bb4725b14ee4b38d93632316f9a",
                "parameters": [
                    {
                        "name": "championPrediction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Predicción de campeón eliminada"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Predicción no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/matches": {
            "get": {
                "summary": "Lista partidos (admin: todos, user: próximos)",
                "operationId": "c4c07f0162b623add6206bfe95f9b37c",
                "responses": {
                    "200": {
                        "description": "Lista de partidos"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Crear partido (solo admin)",
                "operationId": "279f1679408a5ad1a87830c865e4c9b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "home_team_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "away_team_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "phase": {
                                        "type": "string",
                                        "example": "groups"
                                    },
                                    "match_date_time": {
                                        "type": "string",
                                        "example": "2026-06-22 23:00:00"
                                    },
                                    "final_home_goals": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "final_away_goals": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Partido creado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/matches/{matchGame}": {
            "put": {
                "summary": "Actualizar partido (solo admin)",
                "operationId": "56ce041f4f42584c85f61eb2cefa14a7",
                "parameters": [
                    {
                        "name": "matchGame",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "home_team_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "away_team_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "phase": {
                                        "type": "string",
                                        "example": "groups"
                                    },
                                    "match_date_time": {
                                        "type": "string",
                                        "example": "2026-06-22 23:00:00"
                                    },
                                    "final_home_goals": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "final_away_goals": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Partido actualizado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Partido no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Eliminar partido (solo admin)",
                "operationId": "7a87d6186878e7ac01b8fccc16f20bf4",
                "parameters": [
                    {
                        "name": "matchGame",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Partido eliminado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Partido no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}